success

+++ src/events/guildMemberUpdate.js
@@ -8,9 +8,7 @@ module.exports = (client) => {
 
     const dbGuild = await db.guildRepo.getGuild(newMember.guild.id);
 
-    if (dbGuild.roles.muted === null || !oldMember.roles.has(dbGuild.roles.muted)) {
-      return;
-    } else if (!newMember.roles.has(dbGuild.roles.muted) && await db.muteRepo.anyMute(newMember.id, newMember.guild.id)) {
+    if (dbGuild.roles.muted !== null && oldMember.roles.has(dbGuild.roles.muted) && !newMember.roles.has(dbGuild.roles.muted) && await db.muteRepo.anyMute(newMember.id, newMember.guild.id)) {
       return db.muteRepo.deleteMute(newMember.id, newMember.guild.id);
     }
   });
